add/remove tags accordingly; hide empty tag counts, update user name/avatar

+++ posts/views.py
@@ -50,12 +50,20 @@ def edit(request, post_id):
 
 def delete(request, post_id):
     """
-    delete an existing post
+    delete an existing post, and tag count
     """    
     if check_key(request):
         try:
             post = Post.objects(author=request.session['profile'], id=post_id).first()
-            post.delete()
+            if post:
+                for tag in post.tags:
+                    tag_exist = Tag.objects(name=tag.lower()).first()
+                    if tag_exist:
+                        total_count = tag_exist.total_count - 1
+                        tag_exist.total_count = total_count
+                        tag_exist.save()
+
+                post.delete()
         except:
             return HttpResponseRedirect('/yours')
     return HttpResponseRedirect('/yours')
@@ -174,7 +182,7 @@ def tag_search(request):
     """
     tag search
     """
-    tags = Tag.objects[:100]
+    tags = Tag.objects(total_count__gt=0)[:200]
     
     if check_key(request):
         user = get_api(request)
@@ -185,21 +193,3 @@ def tag_search(request):
         'tags' : tags,
         'user' : user,
         }, context_instance=RequestContext(request))
-
-def migrate_tags(request):
-    """
-    migrate tags to tag model
-    """
-    posts = Post.objects()
-    for post in posts:
-        tags = post.tags
-        for tag in tags:
-            tag_exist = Tag.objects(name=tag.lower()).first()
-
-            if tag_exist is None:
-                Tag.objects.create(name=tag.lower(), total_count=1)
-            else:
-                total_count = tag_exist.total_count + 1
-                tag_exist.total_count = total_count
-                tag_exist.save()
-    return HttpResponseRedirect('/yours')
